started moving to float deltatime
- everythine else uses floats, so its easier to just use floats instead
of casting everything
- still some theings that need to be switched over

+++ entity/Entity.cs
@@ -4,7 +4,7 @@ public abstract class Entity {
 
 	public int UpdateIndex { get; protected set; } = 1;
 	public bool shouldRemove = false;
-	public virtual void Update(double updateTime) { }
+	public virtual void Update(float updateTime) { }
 	public virtual void Draw(GameCamera cam) { }
 	public virtual void OnRemove() { }
 	public virtual void OnAdd() { }
